Move CursorManager back to top-level package to have different packages for redis and jdbc, not limited to the CursorManager.
