+++ src/test/java/org/zalando/fahrschein/AbstractCursorManagerTest.java
@@ -4,7 +4,6 @@ import com.google.common.collect.Iterables;
 import com.google.common.collect.Ordering;
 import org.junit.Test;
 import org.springframework.transaction.annotation.Transactional;
-import org.zalando.fahrschein.CursorManager;
 import org.zalando.fahrschein.domain.Cursor;
 
 import java.io.IOException;
