+++ src/test/java/org/zalando/fahrschein/jdbc/PersistentCursorManagerIT.java
@@ -6,6 +6,7 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.transaction.TransactionConfiguration;
 import org.springframework.transaction.annotation.Transactional;
+import org.zalando.fahrschein.AbstractCursorManagerTest;
 import org.zalando.fahrschein.CursorManager;
 import org.zalando.fahrschein.LocalPostgresConfiguration;
